src/main/webapp/app/entities/discussion/update/discussion-update.component.ts
import { Component, inject, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import SharedModule from 'app/shared/shared.module';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';

import { AlertError } from 'app/shared/alert/alert-error.model';
import { EventManager, EventWithContent } from 'app/core/util/event-manager.service';
import { DataUtils, FileLoadError } from 'app/core/util/data-util.service';
import { ICourse } from 'app/entities/course/course.model';
import { CourseService } from 'app/entities/course/service/course.service';
import { DiscussionService } from '../service/discussion.service';
import { IDiscussion } from '../discussion.model';
import { DiscussionFormService, DiscussionFormGroup } from './discussion-form.service';

@Component({
  standalone: true,
  selector: 'jhi-discussion-update',
  templateUrl: './discussion-update.component.html',
  imports: [SharedModule, FormsModule, ReactiveFormsModule],
})
export class DiscussionUpdateComponent implements OnInit {
  isSaving = false;
  discussion: IDiscussion | null = null;

  coursesSharedCollection: ICourse[] = [];

  protected dataUtils = inject(DataUtils);
  protected eventManager = inject(EventManager);
  protected discussionService = inject(DiscussionService);
  protected discussionFormService = inject(DiscussionFormService);
  protected courseService = inject(CourseService);
  protected activatedRoute = inject(ActivatedRoute);

  // eslint-disable-next-line @typescript-eslint/member-ordering
  editForm: DiscussionFormGroup = this.discussionFormService.createDiscussionFormGroup();

  compareCourse = (o1: ICourse | null, o2: ICourse | null): boolean => this.courseService.compareCourse(o1, o2);

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ discussion }) => {
      this.discussion = discussion;
      if (discussion) {
        this.updateForm(discussion);
      }

      this.loadRelationshipsOptions();
    });
  }

  byteSize(base64String: string): string {
    return this.dataUtils.byteSize(base64String);
  }

  openFile(base64String: string, contentType: string | null | undefined): void {
    this.dataUtils.openFile(base64String, contentType);
  }

  setFileData(event: Event, field: string, isImage: boolean): void {
    this.dataUtils.loadFileToForm(event, this.editForm, field, isImage).subscribe({
      error: (err: FileLoadError) =>
        this.eventManager.broadcast(new EventWithContent<AlertError>('lmsOnCloudApp.error', { ...err, key: 'error.file.' + err.key })),
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const discussion = this.discussionFormService.getDiscussion(this.editForm);
    if (discussion.id !== null) {
      this.subscribeToSaveResponse(this.discussionService.update(discussion));
    } else {
      this.subscribeToSaveResponse(this.discussionService.create(discussion));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IDiscussion>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(discussion: IDiscussion): void {
    this.discussion = discussion;
    this.discussionFormService.resetForm(this.editForm, discussion);

    this.coursesSharedCollection = this.courseService.addCourseToCollectionIfMissing<ICourse>(
      this.coursesSharedCollection,
      discussion.course,
    );
  }

  protected loadRelationshipsOptions(): void {
    this.courseService
      .query()
      .pipe(map((res: HttpResponse<ICourse[]>) => res.body ?? []))
      .pipe(map((courses: ICourse[]) => this.courseService.addCourseToCollectionIfMissing<ICourse>(courses, this.discussion?.course)))
      .subscribe((courses: ICourse[]) => (this.coursesSharedCollection = courses));
  }
}
